i dont know how this packet was even working but it was and this makes no sense that this change didnt fix anything but it needed changing anyways so i committed it cause it didnt break anything including the buiold and is correct

+++ src/map/packets/party_member_update.cpp
@@ -67,7 +67,7 @@ CPartyMemberUpdatePacket::CPartyMemberUpdatePacket(CCharEntity* PChar, uint8 Mem
 
     memcpy(data + (0x26), PChar->GetName(), PChar->name.size());
 }
-CPartyMemberUpdatePacket::CPartyMemberUpdatePacket(uint32 id, const int8* name, uint16 memberFlags, uint8 MemberNumber, uint8 ZoneID)
+CPartyMemberUpdatePacket::CPartyMemberUpdatePacket(uint32 id, const int8* name, uint16 memberFlags, uint8 MemberNumber, uint16 ZoneID)
 {
 
     this->type = 0xDD;
